dbs hdn response client

+++ source/common/common-parsers/uk/nhs/hdn/common/parsers/CouldNotParseException.java
@@ -24,12 +24,12 @@ import static java.util.Locale.ENGLISH;
 
 public final class CouldNotParseException extends Exception
 {
-	public CouldNotParseException(final int offset, @NonNls @NotNull final String because)
+	public CouldNotParseException(final long offset, @NonNls @NotNull final String because)
 	{
 		super(format(ENGLISH, "Could not parse at offset %1$s because %2$s", offset, because));
 	}
 
-	public CouldNotParseException(final int offset, @NotNull final Exception cause)
+	public CouldNotParseException(final long offset, @NotNull final Exception cause)
 	{
 		super(format(ENGLISH, "Could not parse at offset %1$s because of exception %2$s", offset, cause.getMessage()), cause);
 	}
